backend/requirements.txt
# Core
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0
pydantic-settings>=2.2.0

# AI/LLM
openai>=1.12.0
langchain>=0.1.10
langchain-community>=0.0.25
langchain-openai>=0.0.8

# Vector Store
faiss-cpu>=1.8.0
numpy>=1.26.0

# Document Parsing
python-docx>=1.1.0
pymupdf>=1.23.25

# Utils
jinja2>=3.1.3
httpx>=0.27.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
